[MED-1002] Fix code review issues and add doctor-user relation

+++ src/main/java/com/domaszekkk/medicalclinic/service/FacilityService.java
@@ -12,7 +12,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 
 @RequiredArgsConstructor
 @Service
@@ -46,6 +45,9 @@ public class FacilityService {
 
     @Transactional
     public void deleteFacility(Long id) {
+        if (!facilityJpaRepository.existsById(id)) {
+            throw new FacilityNotFoundException(id);
+        }
         facilityJpaRepository.deleteById(id);
     }
 }
